+++ src/stringutil.h
@@ -1,6 +1,6 @@
 // stringutil.h -- string utilities.
 //
-// $Id: stringutil.h,v 1.40 2003-01-14 17:52:36 sdennis Exp $
+// $Id: stringutil.h,v 1.41 2003-01-31 06:44:25 sdennis Exp $
 //
 // MUX 2.1
 // Portions are derived from MUX 1.6. Portions are original work.
@@ -108,7 +108,7 @@ void ANSI_String_Skip(struct ANSI_In_Context *pacIn, int maxVisualWidth, int *pn
 void ANSI_String_Copy(struct ANSI_Out_Context *pacOut, struct ANSI_In_Context *pacIn, int nField, int vwMax);
 int ANSI_String_Finalize(struct ANSI_Out_Context *pacOut, int *pnVisualWidth);
 int ANSI_TruncateToField(const char *szString, int nField, char *pField, int maxVisual, int *nVisualWidth, int iEndGoal);
-extern char *strip_ansi(const char *szString, unsigned int *pnString = 0);
+extern char *strip_ansi(const char *szString, size_t *pnString = 0);
 extern char *normal_to_white(const char *);
 extern char *munge_space(char *);
 extern char *trim_spaces(char *);
